ステージ時間制限判定追加

+++ GGJ2019_GirlProject/Assets/Stage/Scripts/StageData.cs
@@ -23,7 +23,7 @@ public class StageData : ScriptableObject
 
 	[SerializeField]
 	[Tooltip("タイムリミット")]
-	public int timeLimit;
+	public float timeLimit;
 
 	[SerializeField]
 	[Tooltip("ギミック1のprefab")]
